requirements.txt
requests
google-cloud-storage
google-cloud-bigquery
psycopg2-binary
dbt-bigquery